// packages/perp/src/typechain/Observer.ts
export type PositionStructOutput = [BigNumber, BigNumber, BigNumber, BigNumber, BigNumber] & {
  balance: BigNumber;
  size: BigNumber;
  entryNotional: BigNumber;
  entrySocialLossIndex: BigNumber;
  entryFundingIndex: BigNumber;
};

export type OrderStruct = {
  balance: PromiseOrValue<BigNumberish>;
  size: PromiseOrValue<BigNumberish>;
};

export type OrderStructOutput = [BigNumber, BigNumber] & {
  balance: BigNumber;
  size: BigNumber;
};

export type RangeStruct = {
  liquidity: PromiseOrValue<BigNumberish>;
  entryFeeIndex: PromiseOrValue<BigNumberish>;
  balance: PromiseOrValue<BigNumberish>;
  sqrtEntryPX96: PromiseOrValue<BigNumberish>;
};

export type RangeStructOutput = [BigNumber, BigNumber, BigNumber, BigNumber] & {
  liquidity: BigNumber;
  entryFeeIndex: BigNumber;
  balance: BigNumber;
  sqrtEntryPX96: BigNumber;
};

export type PortfolioStruct = {
  oids: PromiseOrValue<BigNumberish>[];
  rids: PromiseOrValue<BigNumberish>[];
  position: PositionStruct;
  orders: OrderStruct[];
  ranges: RangeStruct[];
  ordersTaken: PromiseOrValue<BigNumberish>[];
};

export type PortfolioStructOutput = [
  number[],
  number[],
  PositionStructOutput,
  OrderStructOutput[],
  RangeStructOutput[],
  BigNumber[],
] & {
  oids: number[];
  rids: number[];
  position: PositionStructOutput;
  orders: OrderStructOutput[];
  ranges: RangeStructOutput[];
  ordersTaken: BigNumber[];
};

export type BlockInfoStruct = {
  timestamp: PromiseOrValue<BigNumberish>;
  height: PromiseOrValue<BigNumberish>;
};

export type BlockInfoStructOutput = [number, number] & {
  timestamp: number;
  height: number;
};